+++ src/DebouncedButton.h
@@ -62,7 +62,7 @@ public:
      * \return length of last tap in ms if the button has been tapped, else 0.
      *
      */
-    unsigned long tapped();
+    uint32_t tapped();
 
     /*! Test if held on for extended period
      *
@@ -91,14 +91,14 @@ public:
 private:
     uint8_t _threshold;
     uint8_t _delay;
-    unsigned long _lastUpdate;
+    uint32_t _lastUpdate;
     uint8_t _counter;
     bool _state;
-    unsigned long _lastStateChange;
+    uint32_t _lastStateChange;
     bool _pushed;
     uint16_t _repeatCount;
-    unsigned long _lastRepeat;
-    unsigned long _lastOnDuration;
+    uint32_t _lastRepeat;
+    uint32_t _lastOnDuration;
 
 };
